--- src/main/java/nextstep/subway/applicaion/dto/LineRequest.java
@@ -1,14 +1,0 @@
-package nextstep.subway.applicaion.dto;
-
-public class LineRequest {
-    private String name;
-    private String color;
-
-    public String getName() {
-        return name;
-    }
-
-    public String getColor() {
-        return color;
-    }
-}
