--- src/main/java/nextstep/subway/applicaion/dto/LineResponse.java
@@ -1,97 +1,0 @@
-package nextstep.subway.applicaion.dto;
-
-import java.time.LocalDateTime;
-
-import nextstep.subway.domain.Line;
-
-public class LineResponse {
-    private final Long id;
-    private final String name;
-    private final String color;
-    private final LocalDateTime createdDate;
-    private final LocalDateTime modifiedDate;
-
-    private LineResponse(Long id, String name, String color, LocalDateTime createdDate, LocalDateTime modifiedDate) {
-        this.id = id;
-        this.name = name;
-        this.color = color;
-        this.createdDate = createdDate;
-        this.modifiedDate = modifiedDate;
-    }
-
-    public static Builder builder() {
-        return new Builder();
-    }
-
-    public static LineResponse from(Line line) {
-        return builder()
-            .id(line.getId())
-            .name(line.getName())
-            .color(line.getColor())
-            .createdDate(line.getCreatedDate())
-            .modifiedDate(line.getModifiedDate())
-            .build();
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getColor() {
-        return color;
-    }
-
-    public LocalDateTime getCreatedDate() {
-        return createdDate;
-    }
-
-    public LocalDateTime getModifiedDate() {
-        return modifiedDate;
-    }
-
-    public static class Builder {
-        private Long id;
-        private String name;
-        private String color;
-        private LocalDateTime createdDate;
-        private LocalDateTime modifiedDate;
-
-        private Builder() {
-        }
-
-        public Builder id(Long id) {
-            this.id = id;
-            return this;
-        }
-
-        public Builder name(String name) {
-            this.name = name;
-            return this;
-        }
-
-        public Builder color(String color) {
-            this.color = color;
-            return this;
-        }
-
-        public Builder createdDate(LocalDateTime createdDate) {
-            this.createdDate = createdDate;
-            return this;
-        }
-
-        public Builder modifiedDate(LocalDateTime modifiedDate) {
-            this.modifiedDate = modifiedDate;
-            return this;
-        }
-
-        public LineResponse build() {
-            return new LineResponse(
-                id, name, color, createdDate, modifiedDate
-            );
-        }
-    }
-}
