+++ src/main/java/nextstep/subway/common/exception/DuplicateColumnException.java
@@ -1,0 +1,7 @@
+package nextstep.subway.common.exception;
+
+public class DuplicateColumnException extends RuntimeException {
+    public DuplicateColumnException(ColumnName columnName) {
+        super(ErrorMessage.DUPLICATE_COLUMN.getMessage(columnName.getName()));
+    }
+}
