+++ src/main/java/nextstep/subway/common/exception/EntityNotFoundException.java
@@ -1,0 +1,7 @@
+package nextstep.subway.common.exception;
+
+public class EntityNotFoundException extends RuntimeException {
+    public EntityNotFoundException() {
+        super(ErrorMessage.ENTITY_NOT_FOUND.getMessage());
+    }
+}
