--- src/main/java/nextstep/subway/domain/Line.java
@@ -1,48 +1,0 @@
-package nextstep.subway.domain;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Entity
-public class Line extends BaseEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(unique = true)
-    private String name;
-
-    private String color;
-
-    public Line() {
-    }
-
-    public Line(String name, String color) {
-        this.name = name;
-        this.color = color;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getColor() {
-        return color;
-    }
-
-    public void edit(String name, String color) {
-        this.name = name;
-        this.color = color;
-    }
-
-    public boolean notMatchName(String name) {
-        return !this.name.equals(name);
-    }
-}
