--- src/main/java/nextstep/subway/exception/DuplicateColumnException.java
@@ -1,7 +1,0 @@
-package nextstep.subway.exception;
-
-public class DuplicateColumnException extends RuntimeException {
-    public DuplicateColumnException(ColumnName columnName) {
-        super(ErrorMessage.DUPLICATE_COLUMN.getMessage(columnName.getName()));
-    }
-}
