--- src/main/java/nextstep/subway/exception/EntityNotFoundException.java
@@ -1,7 +1,0 @@
-package nextstep.subway.exception;
-
-public class EntityNotFoundException extends RuntimeException {
-    public EntityNotFoundException() {
-        super(ErrorMessage.ENTITY_NOT_FOUND.getMessage());
-    }
-}
