+++ src/main/java/nextstep/subway/line/domain/dto/LineRequest.java
@@ -1,0 +1,14 @@
+package nextstep.subway.line.domain.dto;
+
+public class LineRequest {
+    private String name;
+    private String color;
+
+    public String getName() {
+        return name;
+    }
+
+    public String getColor() {
+        return color;
+    }
+}
