+++ src/main/java/nextstep/subway/line/domain/model/Line.java
@@ -1,0 +1,50 @@
+package nextstep.subway.line.domain.model;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+import nextstep.subway.common.domain.model.BaseEntity;
+
+@Entity
+public class Line extends BaseEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(unique = true)
+    private String name;
+
+    private String color;
+
+    public Line() {
+    }
+
+    public Line(String name, String color) {
+        this.name = name;
+        this.color = color;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getColor() {
+        return color;
+    }
+
+    public void edit(String name, String color) {
+        this.name = name;
+        this.color = color;
+    }
+
+    public boolean notMatchName(String name) {
+        return !this.name.equals(name);
+    }
+}
