+++ src/main/java/nextstep/subway/line/domain/repository/LineRepository.java
@@ -1,0 +1,9 @@
+package nextstep.subway.line.domain.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import nextstep.subway.line.domain.model.Line;
+
+public interface LineRepository extends JpaRepository<Line, Long> {
+    boolean existsByName(String name);
+}
