+++ src/main/java/nextstep/subway/station/domain/dto/StationRequest.java
@@ -1,0 +1,9 @@
+package nextstep.subway.station.domain.dto;
+
+public class StationRequest {
+    private String name;
+
+    public String getName() {
+        return name;
+    }
+}
