+++ src/main/java/nextstep/subway/station/ui/StationController.java
@@ -1,0 +1,37 @@
+package nextstep.subway.station.ui;
+
+import nextstep.subway.station.applicaion.StationService;
+import nextstep.subway.station.domain.dto.StationRequest;
+import nextstep.subway.station.domain.dto.StationResponse;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.net.URI;
+import java.util.List;
+
+@RestController
+public class StationController {
+    private final StationService stationService;
+
+    public StationController(StationService stationService) {
+        this.stationService = stationService;
+    }
+
+    @PostMapping("/stations")
+    public ResponseEntity<StationResponse> createStation(@RequestBody StationRequest stationRequest) {
+        StationResponse station = stationService.saveStation(stationRequest);
+        return ResponseEntity.created(URI.create("/stations/" + station.getId())).body(station);
+    }
+
+    @GetMapping(value = "/stations", produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<List<StationResponse>> showStations() {
+        return ResponseEntity.ok().body(stationService.findAllStations());
+    }
+
+    @DeleteMapping("/stations/{id}")
+    public ResponseEntity<Void> deleteStation(@PathVariable Long id) {
+        stationService.deleteStationById(id);
+        return ResponseEntity.noContent().build();
+    }
+}
